Move admin merge tool views to separate package
