+++ datahub/company/admin/merge/constants.py
@@ -1,0 +1,1 @@
+MERGE_COMPANY_TOOL_FEATURE_FLAG = 'admin-merge-company-tool'
